requirements.txt
requests
pydantic
google-generativeai
google-genai
google-api-core
pdf2image
Pillow
polars
python-fasthtml